+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+### 1.5.0
+- Added FREObject.AsDateTime()
+- Added DateTime.ToFREObject()
+- Refactor FrePointSharp
+- Refactor FreRectangleSharp
+- Added Color.ToFREObject()
+- Added FREObject.AsColor()
+- Added int, double, boolean, string conversions to FREArray
+
 ### 1.4.0
 Allow GetAsString to return null strings
 Updated ANEUtils
@@ -7,6 +16,6 @@ Updated ANEUtils
 - Added AsColor() method
 - Added GetScaleFactor() method to WinAPI
 - Added OnFinalise() required method
-- **C++ project now needs reference to ..\packages\TuaRua.FreSharp.1.3.0\FreSharp\$(PlatformTarget)\FreSharp.dll**
+- **C++ project now needs reference to ..\packages\TuaRua.FreSharp.1.5.0\FreSharp\$(PlatformTarget)\FreSharp.dll**
 - Refactor